+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#V 0.3
+* bug fixing
+* now the analysis is perfomed using n_networks number of graphs generated independently form the initial one
+* added a monitoring function for the underlying markov chain
+* added an external function computing the similarity between two consistent graph representations
+
 #V 0.2
 * bug fixing
 * added little documentation